BJS2-15571 - Повышенная сложность

+++ src/main/java/school/faang/user_service/dto/PaymentRequest.java
@@ -2,12 +2,10 @@ package school.faang.user_service.dto;
 
 import lombok.Builder;
 
-import java.math.BigDecimal;
-
 @Builder
 public record PaymentRequest(
     long paymentNumber,
-    BigDecimal amount,
+    double amount,
     Currency currency
 ) {
 }
